gitignore

+++ backend/main/resources/usuario.py
@@ -7,6 +7,7 @@ from main.models.reseñas import usuario_reseñas
 from sqlalchemy import func, desc
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from main.auth.decorators import role_required
+from sqlalchemy.exc import IntegrityError
 
 app = Flask(__name__)
 api = Api(app)
@@ -24,22 +25,81 @@ class Usuario(Resource):
             return usuario.to_json()
 
     # Eliminar recurso usuario
-    @role_required(roles=["admin"])
+
+    @role_required(roles=["admin", "bibliotecario"])
     def delete(self, id):
         usuario = db.session.query(UsuariosModel).get_or_404(id)
+
+        if usuario.rol in ["admin", "bibliotecario"]:
+            abort(403, message="No se puede eliminar a otros administradores o bibliotecarios")
+
         db.session.delete(usuario)
         db.session.commit()
-        return ' Eliminado correctamente ', 204
-    
+        return {'mensaje': 'Eliminado correctamente', 'usuario_eliminado': usuario.to_json()}, 200
+        
+
     # Modificar el recurso usuario
-    @role_required(roles=["admin", "bibliotecario"])
+    @jwt_required()
     def put(self, id):
         usuario = db.session.query(UsuariosModel).get_or_404(id)
+
+        if usuario.rol in ["admin", "bibliotecario"]:
+            abort(403, message="No se puede editar a otros administradores o bibliotecarios")
+
         data = request.get_json()
         for key, value in data.items():
             setattr(usuario, key, value)
         db.session.commit()
-        return {'mensaje': 'El usuario ha sido editado con éxito', 'usuario_modificado': usuario.to_json()}, 201
+        return {'mensaje': 'Editado correctamente', 'usuario_editaro': usuario.to_json()}, 200
+
+    """ if current_user.rol not in ["admin", "bibliotecario"]:
+                abort(403, message="No tienes permiso para editar otros usuarios")
+    current_user = UsuariosModel.query.get(current_user_id)#dsps agregar         current_user_id = get_jwt_identity()
+        if usuario.rol in ["admin", "bibliotecario"]:
+                abort(403, message="No se puede editar a otros administradores o bibliotecarios")
+
+        if current_user_id != str(usuario.id):
+            if usuario.rol in ["alumno", "users"]:
+                abort(403, message="No tienes permiso para editar otros usuarios")
+
+        data = request.get_json()
+        for key, value in data.items():
+            setattr(usuario, key, value)
+        
+        try:
+            db.session.commit()
+        except IntegrityError as e:
+            db.session.rollback()
+            if "UNIQUE constraint failed" in str(e):
+                abort(400, message="El correo electrónico ya está en uso")
+            else:
+                abort(500, message="Error interno del servidor")
+
+        return {'mensaje': 'El usuario ha sido editado con éxito', 'usuario_modificado': usuario.to_json()}, 201 """
+    
+    @role_required(roles=["admin", "bibliotecario"])
+    @jwt_required()
+    def update_usuario(self, usuario_id):
+        usuario = db.session.query(UsuariosModel).get_or_404(usuario_id)
+        data = request.get_json()
+
+        if 'rol' in data and data['rol'] == "alumno":
+            usuario.rol = "alumno"
+
+        for key, value in data.items():
+            setattr(usuario, key, value)
+        
+        try:
+            db.session.commit()
+        except IntegrityError as e:
+            db.session.rollback()
+            if "UNIQUE constraint failed" in str(e):
+                abort(400, message="El correo electrónico ya está en uso")
+            else:
+                abort(500, message="Error interno del servidor")
+        
+        return {'mensaje': 'El usuario ha sido actualizado con éxito', 'usuario_modificado': usuario.to_json()}, 200
+
 
 class Usuarios(Resource): 
     @role_required(roles=["admin", "bibliotecario"])
